Fix timeout issues on adapter open
Previously it was necessary to add waiting between closing and opening
an adapter or else pc-ble-driver could crash or the adapter did not respond.

This commit fix that by upgrading to the latest version of Boost
and by controlling how the destruction of objects are done in
uart_boost destructor.

In addition a wait is added between the opening of the adapter and
making it available to the upper layers. That seems to fix most timeout
issues (tested with PCA10028, PCA10030, PCA10040).