Remove boost dependency
The boost library is large and only the serial port
part is used. This commit removes the boost
dependency by using ASIO instead of pulling
in large portions of the boost library.

To get the ASIO library the conan package manager is used.

This commit is one of two; this one prepares the
dependency of ASIO through conan. The next one
contains the necessary changes to the source code.

It is done this way because we may not end up using
the conan package manager.