Remove boost dependency
The boost library is large and only the serial port
part is used. This commit removes the boost
dependency by using ASIO instead of pulling
in large portions of the boost library.

This commit is one of two; this one contains
the necessary changes to the source code to use
ASIO standalone. The other one prepares the
dependency of ASIO through conan.